refactor(eval): remove the pre-IR compile_rule_legacy path
Now that compile_rule routes exclusively through the IR
(lower_rule -> compile_to_compiled) and has shipped as the only path,
drop the dual-path scaffolding:

- remove compile_rule_legacy and its private validate_condition_refs
  helper (compile_detection and yaml_to_json_map stay; used by the engine
  filter path and correlation compiler).
- remove the compile_rule_legacy re-export and doc references.
- replace crates/rsigma-ir/tests/differential.rs (IR vs legacy parity)
  with ir_eval_semantics.rs, which asserts the same corners (vacuous
  all-of, them skipping _-names, contains/cidr/keywords, include_event,
  modifier contradictions) directly against the IR-compiled result.

The rule_load dual-path bench comparison is moot with no second path to
compare against.
