+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The engine's per-field substring bloom pre-filter seeded its hasher from ahash's
 ### Intermediate representation crate and IR-backed compile (#360, #362, #363, #364, #368)
 
 - **`rsigma-ir`** — new sync-only crate with HIR types (`IrRule`, `IrDetection`, `IrMatcher`, `IrCondition`, `IrCorrelation`, `IrFilter`) and `lower_rule` / `lower_*` that resolve modifiers before eval. Quantified selectors are preserved so evaluation stays count-based. The matcher model is faithful and lossless: string matches keep a wildcard-aware, original-case `IrPattern` and encoding modifiers stay explicit as `IrEncoding` steps, so lowering never lowercases, compiles regexes, or expands encodings.
-- **`rsigma-eval`** — `compile_rule` routes through `lower_rule` → `compile_to_compiled`, which performs the physical work (lowercasing, regex/aho-corasick, encoding expansion). `compile_rule_legacy` remains for dual-path differential tests. Public physical API (`CompiledRule`, `evaluate_rule`, `Engine`) is unchanged.
+- **`rsigma-eval`** — `compile_rule` routes through `lower_rule` → `compile_to_compiled`, which performs the physical work (lowercasing, regex/aho-corasick, encoding expansion). Public physical API (`CompiledRule`, `evaluate_rule`, `Engine`) is unchanged.
 - **`rsigma-ir` HIR cache** — all HIR types derive `serde::{Serialize, Deserialize}` (and the embedded `rsigma-parser` types gained `Deserialize`), so a lowered rule set round-trips. A new `cache` module serializes a slice of `IrRule` to a versioned, self-describing blob (a `HirCacheHeader` with `HIR_SCHEMA_VERSION` plus the rules) via `encode_rules` / `decode_rules`, with `to_json` for a debug export. `decode_rules` version-checks the header before decoding the rules. CBOR is the binary format because the embedded `LogSource`'s `#[serde(flatten)]` map has an unknown length that fixed-layout encoders reject.
 - **`rsigma-eval` restart cache** — `Engine::save_hir` / `Engine::load_hir` persist and reload the engine's lowered rules through the HIR cache, so a warm start skips parse, pipeline, and lowering. The engine retains the post-pipeline HIR for rules added via the parsed-rule paths; filter injections and precompiled rules are not captured, so re-apply filters after `load_hir`.
 - **`rsigma-ir`** — opt-in, semantics-preserving HIR optimization passes in `optimize`: `flatten_condition` (merge nested same-kind boolean groups, collapse `Not(Not)`, unwrap single-child groups, drop idempotent duplicate siblings), `eliminate_dead_detections` (prune detections no condition can reference, honoring `them`/glob selector patterns and recursing into `Conditional`), and `common_subexpressions` (a non-mutating report of repeated detection items). `optimize_rule` runs the structural passes in order. The passes are not run by the default eval or convert paths, so compiled-matcher behavior and byte-identical backend output are unchanged; a differential test confirms each pass preserves the match decision and the set of matched selections and fields.
